pyproject.toml
[tool.poetry]
name = "cdk-aws-lambda-internal-extension"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "^2.87.0"
aws-cdk-aws-lambda-python-alpha = "^2.87.0a0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
force_grid_wrap = 2
combine_as_imports = true
lines_between_sections = 0

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
